Move the start up logic into cmd pkg:
This makes it more straight forward to
handle exiting cleanly.
